// Frontend/Módulo - Consultas Amsys Navega/consulta-pagos-amsys-navega/src/app/models/NavegaPayment.ts
export interface NavegaPayment {
    id?: number;
    navegaCode: string;
    productCode: number;
    currency: string;
    paymentAmount: number;
    bank: number;
    bankAuthorization: number;
    exchangeRate: number;
    paymentDate: string;
    idDeaOrganization: number;
    ebsAccount: string;
    transactionSts: string;
}